chore: Remove Rust API calls from outside Dart services
