+++ mobile/lib/common/settings_screen.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 
 import 'package:f_logs/f_logs.dart';
 import 'package:flutter/material.dart';
+import 'package:get_10101/common/application/channel_info_service.dart';
 import 'package:get_10101/features/trade/settings_screen.dart';
 import 'package:get_10101/features/wallet/settings_screen.dart';
 import 'package:get_10101/bridge_generated/bridge_definitions.dart' as bridge;
@@ -11,7 +12,6 @@ import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:share_plus/share_plus.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-import 'package:get_10101/ffi.dart' as rust;
 
 class SettingsScreen extends StatefulWidget {
   const SettingsScreen({required this.fromRoute, super.key});
@@ -29,8 +29,6 @@ class _SettingsScreenState extends State<SettingsScreen> {
 
   @override
   void initState() {
-    var nodeId = rust.api.getNodeId();
-    _nodeId = nodeId;
     loadValues();
     super.initState();
   }
@@ -52,6 +50,10 @@ class _SettingsScreenState extends State<SettingsScreen> {
     String commit = const String.fromEnvironment('COMMIT', defaultValue: 'not available');
     String branch = const String.fromEnvironment('BRANCH', defaultValue: 'not available');
 
+    final channelService = context.read<ChannelInfoService>();
+
+    _nodeId = channelService.getNodeId();
+
     return Scaffold(
       appBar: AppBar(title: const Text("Settings")),
       body: SafeArea(
@@ -69,12 +71,12 @@ class _SettingsScreenState extends State<SettingsScreen> {
             children: [
               ElevatedButton(
                   onPressed: () {
-                    rust.api.closeChannel();
+                    channelService.closeChannel();
                   },
                   child: const Text("Close channel")),
               ElevatedButton(
                   onPressed: () {
-                    rust.api.forceCloseChannel();
+                    channelService.forceCloseChannel();
                   },
                   child: const Text("Force-close channel")),
             ],
